Update import paths
