+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Auth from './components/Auth';
+import Auth from './components/Auth/Auth';
 import Login from './components/Auth/Login';
 import Signup from './components/Auth/Signup';
 
@@ -10,9 +10,9 @@ function App() {
     <div className="App">
       <Router>
         <Routes>
-          <Route exact path='/' element={<Auth />}/>
-          <Route exact path="/login" element={<Login /> }/>
-          <Route exact path="/signup" element={<Signup /> }/>          
+          <Route exact path="/" element={<Auth />} />
+          <Route exact path="/login" element={<Login />} />
+          <Route exact path="/signup" element={<Signup />} />
         </Routes>
       </Router>
     </div>
