+++ src/components/Auth/Login.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Login = () => (
+  <div>
+    <h1>login form</h1>
+  </div>
+);
+
+export default Login;
