+++ src/components/Auth/Signup.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Signup = () => (
+  <div>
+    <h1>signup form</h1>
+  </div>
+);
+
+export default Signup;
