--- src/components/Login.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Login = () => {
-  return (
-    <div>
-      <h1>login form</h1>
-    </div>
-  )
-}
-
-export default Login
