--- src/components/Signup.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Signup = () => {
-  return (
-    <div>
-      <h1>signup form</h1>
-    </div>
-  )
-}
-
-export default Signup
